+++ common-core/src/main/java/cn/featherfly/common/lang/AssertIllegalArgument.java
@@ -31,7 +31,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not null.
      *
-     * @param object    object
+     * @param object object
      * @param arguDescp argument descption
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotNull(java.lang.Object,
      *      java.lang.String)
@@ -43,7 +43,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not blank.
      *
-     * @param text      text
+     * @param text text
      * @param arguDescp argument descption
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotBlank(java.lang.String,
      *      java.lang.String)
@@ -55,7 +55,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not empty.
      *
-     * @param obj       obj
+     * @param obj obj
      * @param arguDescp arguDescp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(java.lang.Object,
      *      java.lang.String)
@@ -67,7 +67,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not empty.
      *
-     * @param text      text
+     * @param text text
      * @param arguDescp argument descption
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(java.lang.String,
      *      java.lang.String)
@@ -79,7 +79,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not empty.
      *
-     * @param array     array
+     * @param array array
      * @param arguDescp arguDescp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(java.lang.Object[],
      *      java.lang.String)
@@ -92,7 +92,7 @@ public final class AssertIllegalArgument {
      * Checks if is not empty.
      *
      * @param collection collection
-     * @param arguDescp  arguDescp
+     * @param arguDescp arguDescp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(java.util.Collection,
      *      java.lang.String)
      */
@@ -103,7 +103,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not empty.
      *
-     * @param map       map
+     * @param map map
      * @param arguDescp arguDescp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(java.util.Map,
      *      java.lang.String)
@@ -113,9 +113,20 @@ public final class AssertIllegalArgument {
     }
 
     /**
+     * Checks if is array.
+     *
+     * @param array the array
+     * @param arguDescp the argu descp
+     * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isArray(Object, String)
+     */
+    public static void isArray(Object array, String arguDescp) {
+        ASSERT.isArray(array, arguDescp);
+    }
+
+    /**
      * Checks if is exists.
      *
-     * @param file      file
+     * @param file file
      * @param arguDescp argument descption
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isExists(java.io.File,
      *      java.lang.String)
@@ -128,7 +139,7 @@ public final class AssertIllegalArgument {
      * Checks if is instance of.
      *
      * @param clazz clazz
-     * @param obj   obj
+     * @param obj obj
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isInstanceOf(java.lang.Class,
      *      java.lang.Object)
      */
@@ -140,7 +151,7 @@ public final class AssertIllegalArgument {
      * Checks if is parent.
      *
      * @param parentType parentType
-     * @param subType    subType
+     * @param subType subType
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isParent(java.lang.Class,
      *      java.lang.Class)
      */
@@ -151,7 +162,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is file.
      *
-     * @param file      the file
+     * @param file the file
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isFile(java.io.File,
      *      java.lang.String)
@@ -163,7 +174,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is directory.
      *
-     * @param file      the file
+     * @param file the file
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isDirectory(java.io.File,
      *      java.lang.String)
@@ -185,10 +196,10 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is in range.
      *
-     * @param <N>       the number type
-     * @param value     the value
-     * @param min       the min
-     * @param max       the max
+     * @param <N> the number type
+     * @param value the value
+     * @param min the min
+     * @param max the max
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isInRange(java.lang.Number,
      *      java.lang.Number, java.lang.Number, java.lang.String)
@@ -200,10 +211,10 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is in range.
      *
-     * @param <N>   the number type
+     * @param <N> the number type
      * @param value the value
-     * @param min   the min
-     * @param max   the max
+     * @param min the min
+     * @param max the max
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isInRange(SerializableNumberSupplier,
      *      java.lang.Number, java.lang.Number)
      */
@@ -214,9 +225,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is gt.
      *
-     * @param <N>       the number type
-     * @param value     the value
-     * @param min       the min
+     * @param <N> the number type
+     * @param value the value
+     * @param min the min
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isGt(java.lang.Number,
      *      java.lang.Number, java.lang.String)
@@ -228,9 +239,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is gt.
      *
-     * @param <N>   the number type
+     * @param <N> the number type
      * @param value the value
-     * @param min   the min
+     * @param min the min
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isGt(SerializableNumberSupplier,
      *      java.lang.Number)
      */
@@ -241,9 +252,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is ge.
      *
-     * @param <N>       the number type
-     * @param value     the value
-     * @param min       the min
+     * @param <N> the number type
+     * @param value the value
+     * @param min the min
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isGe(java.lang.Number,
      *      java.lang.Number, java.lang.String)
@@ -255,9 +266,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is ge.
      *
-     * @param <N>   the number type
+     * @param <N> the number type
      * @param value the value
-     * @param min   the min
+     * @param min the min
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isGe(SerializableNumberSupplier,
      *      java.lang.Number)
      */
@@ -268,9 +279,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is lt.
      *
-     * @param <N>       the number type
-     * @param value     the value
-     * @param max       the max
+     * @param <N> the number type
+     * @param value the value
+     * @param max the max
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isLt(java.lang.Number,
      *      java.lang.Number, java.lang.String)
@@ -282,9 +293,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is lt.
      *
-     * @param <N>   the number type
+     * @param <N> the number type
      * @param value the value
-     * @param max   the max
+     * @param max the max
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isLt(SerializableNumberSupplier,
      *      java.lang.Number)
      */
@@ -295,9 +306,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is le.
      *
-     * @param <N>       the number type
-     * @param value     the value
-     * @param max       the max
+     * @param <N> the number type
+     * @param value the value
+     * @param max the max
      * @param arguDescp the argu descp
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isLe(java.lang.Number,
      *      java.lang.Number, java.lang.String)
@@ -309,9 +320,9 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is le.
      *
-     * @param <N>   the number type
+     * @param <N> the number type
      * @param value the value
-     * @param max   the max
+     * @param max the max
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isLe(SerializableNumberSupplier,
      *      java.lang.Number)
      */
@@ -322,7 +333,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not null.
      *
-     * @param <T>              the generic type
+     * @param <T> the generic type
      * @param propertySupplier the property supplier
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotNull(SerializableSupplier)
      */
@@ -343,7 +354,7 @@ public final class AssertIllegalArgument {
     /**
      * Checks if is not empty.
      *
-     * @param <T>              the generic type
+     * @param <T> the generic type
      * @param propertySupplier the property supplier
      * @see cn.featherfly.common.lang.asserts.LocalizedAssert#isNotEmpty(SerializableSupplier)
      */